requirements.txt
fastapi
uvicorn
easyocr
opencv-python-headless
numpy
pillow
